Fix bugs in model
