+++ language/bert/basic/utils/__init__.py
@@ -1,7 +1,5 @@
 from .dataset import load_data_wiki
 from .bert import BERTModel
-from .timer import Timer
-from .plot import Animator, Accumulator
+from .profile import Timer, Accumulator
 
-__all__ = ['load_data_wiki', 'BERTModel', 'Timer',
-           'Animator', 'Accumulator']
+__all__ = ['load_data_wiki', 'BERTModel', 'Timer', 'Accumulator']
