+++ language/bert/basic/utils/profile.py
@@ -1,0 +1,33 @@
+import time
+import numpy as np
+
+class Timer:
+    def __init__(self):
+        self.times = []
+        self.start()
+
+    def start(self):
+        self.tik = time.time()
+
+    def stop(self):
+        self.times.append(time.time() - self.tik)
+        return self.times[-1]
+
+    def sum(self):
+        return sum(self.times)
+
+    def avg(self):
+        return sum(self.times) / len(self.times)
+
+class Accumulator:
+    def __init__(self, n):
+        self.data = [0.0] * n
+
+    def add(self, *args):
+        self.data = [a + float(b) for a, b in zip(self.data, args)]
+
+    def reset(self):
+        self.data = [0.0] * len(self.data)
+
+    def __getitem__(self, idx):
+        return self.data[idx]
