--- language/bert/basic/utils/timer.py
@@ -1,20 +1,0 @@
-import time
-import numpy as np
-
-class Timer:
-    def __init__(self):
-        self.times = []
-        self.start()
-
-    def start(self):
-        self.tik = time.time()
-
-    def stop(self):
-        self.times.append(time.time() - self.tik)
-        return self.times[-1]
-
-    def sum(self):
-        return sum(self.times)
-
-    def avg(self):
-        return sum(self.times) / len(self.times)
